refactor: 데모데이 관리자 권한 검사기를 서비스 공용 위치로 이동
